@@ -2,14 +2,12 @@
 
 namespace App\Misc;
 
-use Zttp\Zttp;
-
 class WpApi
 {
     const ENDPOINT_MODULES = 'freescout/v1/modules';
 
-    const METHOD_GET = 'get';
-    const METHOD_POST = 'post';
+    const METHOD_GET = 'GET';
+    const METHOD_POST = 'POST';
 
     const ACTION_CHECK_LICENSE = 'check_license';
     const ACTION_ACTIVATE_LICENSE = 'activate_license';
@@ -28,6 +26,23 @@ class WpApi
         }
     }
 
+    public static function httpRequest($method, $url, $params)
+    {
+        $client = new \GuzzleHttp\Client();
+
+        if ($method == self::METHOD_POST) {
+            return $client->request('POST', $url, [
+                'connect_timeout' => 10,
+                'form_params' => $params,
+            ]);
+        } else {
+            return $client->request('GET', $url, [
+                'connect_timeout' => 10,
+                'query' => $params,
+            ]);
+        }
+    }
+
     /**
      * API request.
      */
@@ -35,14 +50,13 @@ class WpApi
     {
         self::$lastError = null;
 
-        $options = ['connect_timeout' => 7]; // seconds
-
         try {
-            $response = Zttp::withOptions($options)->$method(self::url($endpoint, $alternative_api), $params);
+            $response = self::httpRequest($method, self::url($endpoint, $alternative_api), $params);
         } catch (\Exception $e) {
             if (!$alternative_api) {
                 return self::request($method, $endpoint, $params, true);
             }
+            \Helper::logException($e, 'WpApi');
             self::$lastError = [
                 'code'    => $e->getCode(),
                 'message' => $e->getMessage(),
@@ -52,8 +66,9 @@ class WpApi
         }
 
         // https://guzzle3.readthedocs.io/http-client/response.html
-        if ($response->status() < 500) {
-            $json = $response->json();
+        if ($response->getStatusCode() < 500) {
+            $json = \Helper::jsonToArray($response->getBody());
+
             if (!empty($json['code']) && !empty($json['message']) &&
                 !empty($json['data']) && !empty($json['data']['status']) && $json['data']['status'] != 200
             ) {
