@@ -12,7 +12,7 @@ return [
     | or any other location as required by the application or its packages.
     */
 
-    'version' => '1.7.27',
+    'version' => '1.8.6',
 
     /*
     |--------------------------------------------------------------------------
@@ -92,7 +92,7 @@ return [
     */
 
     'locale'          => env('APP_LOCALE', 'en'),
-    'locales'         => ['en', 'hr', 'da', 'nl', 'fr', 'de', 'ja', 'it', 'pl', 'pt-PT', 'pt-BR', 'ru', 'es', 'sk', 'sv','fa'],
+    'locales'         => ['en', 'hr', 'da', 'nl', 'fr', 'de', 'ja', 'it', 'pl', 'pt-PT', 'pt-BR', 'ru', 'es', 'sk', 'sv'],
     'default_locale'  => 'en',
 
     /*
@@ -268,7 +268,7 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | Case insensitive regular expression, containing a list of
+    | Case insensitive regular expression, containing a list of 
     | mail server error responses, returned when a mail server can not deliver an email
     | to one or more recipients. If FreeScout receives one of the listed
     | error responses from the mail server, it does not try to resend the email
